fix(stats): address PR review comments and update docs snapshot
- Fix GpuUtilizationRow type mismatch (count → provider_count)
- Add null checks for findFirstBlockSince instead of unsafe casts
- Replace `as any` with typed casts in stats service spec
- Handle empty query results in getActiveResourcesAtHeight
- Add COALESCE to SUM aggregates in provider dashboard query
- Handle null values from LEFT JOIN in provider earnings reducer
- Use z.number().int() for block height in BME schema
- Use ProviderDashboardService.name in integration test describe
- Use object property shorthand in findClosedLeases
- Update docs snapshot with new BME status-history endpoint

+++ apps/api/src/dashboard/http-schemas/bme-status-history/bme-status-history.schema.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 export const BmeStatusHistoryResponseSchema = z.array(
   z.object({
-    height: z.number().openapi({ example: 12345678 }),
+    height: z.number().int().openapi({ example: 12345678 }),
     date: z.date().openapi({ example: "2024-01-01T00:00:00.000Z" }),
     previousStatus: z.string().openapi({ example: "healthy" }),
     newStatus: z.string().openapi({ example: "warning" }),
